requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
python-dateutil>=2.9.0
schedule>=1.2.0
